% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANN.R
\name{z.Ann10}
\alias{z.Ann10}
\title{ANN correlation}
\usage{
z.Ann10(pres.pr, temp.pr, tolerance, verbose)
}
\arguments{
\item{pres.pr}{pseudo-reduced pressure}

\item{temp.pr}{pseudo-reduced temperature}
}
\description{
ANN correlation
}
