% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANN.R
\name{z.Ann10}
\alias{z.Ann10}
\title{Artificial Neural Network correlation}
\usage{
z.Ann10(pres.pr, temp.pr, tolerance, verbose)
}
\arguments{
\item{pres.pr}{pseudo-reduced pressure}

\item{temp.pr}{pseudo-reduced temperature}

\item{tolerance}{controls the iteration accuracy}

\item{verbose}{print internal}
}
\description{
Artificial Neural Network correlation
}
