% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_tables.R
\name{DPR.genDataset7p4t}
\alias{DPR.genDataset7p4t}
\title{Generate a dataset of z values calculated by DPR}
\usage{
DPR.genDataset7p4t(to_disk = FALSE)
}
\description{
Generate a dataset of z values calculated by DPR
}
