% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex34}
\alias{ex34}
\title{Example 34 fro McCain chapter 03}
\format{An object of class \code{data.frame} with 3 rows and 4 columns.}
\usage{
ex34
}
\description{
Example 34 fro McCain chapter 03
}
\keyword{datasets}
