% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex35}
\alias{ex35}
\title{Example 35 from McCain chapter 03}
\format{An object of class \code{data.frame} with 4 rows and 4 columns.}
\usage{
ex35
}
\description{
Example 35 from McCain chapter 03
}
\keyword{datasets}
