% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Standing-Katz.R
\name{getCurvesDigitized}
\alias{getCurvesDigitized}
\title{Get a numeric vector of digitized curves available by Tpr}
\usage{
getCurvesDigitized(pprRange)
}
\arguments{
\item{pprRange}{Takes one of 4 values: "lp": low pressure, or "hp" for
high pressure; "all": all curves; "common": only curves that are common to hp
and lp}
}
\description{
Get a numeric vector of digitized curves available by Tpr
}
\examples{
getCurvesDigitized(pprRange = "lp")
getCurvesDigitized(pprRange = "common")
}
