% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Standing-Katz.R
\name{getStandingKatzMatrix}
\alias{getStandingKatzMatrix}
\title{Generate a matrix of Standing-Katz pseudo-reduced pressure and tenperarture}
\usage{
getStandingKatzMatrix(ppr_vector, tpr_vector, pprRange = "lp")
}
\arguments{
\item{ppr_vector}{a vector of pseudo-reduced pressure}

\item{tpr_vector}{a vector of pseudo-reduced temperatures}

\item{pprRange}{Takes one of two values: "lp": low pressure, or "hp" for
high pressure}
}
\description{
Generate a matrix of Standing-Katz pseudo-reduced pressure and tenperarture
}
