% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Standing-Katz.R
\name{getStandingKatzTpr}
\alias{getStandingKatzTpr}
\title{Get a numeric vector of the digitized curves available for Pseudo Reduced
Temperature}
\usage{
getStandingKatzTpr(pprRange = NULL)
}
\arguments{
\item{pprRange}{Takes one of 4 values: "lp": low pressure, or "hp" for
high pressure; "all": all curves; "common": only curves that are common to hp
and lp}
}
\description{
Get a numeric vector of the digitized curves available for Pseudo Reduced
Temperature
}
\examples{
getStandingKatzTpr(pprRange = "lp")
getStandingKatzTpr(pprRange = "common")
}
