% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hall-Yarborough.R
\name{z.HallYarborough}
\alias{z.HallYarborough}
\title{Hall-Yarborough correlation}
\usage{
z.HallYarborough(pres.pr, temp.pr, tolerance = 1e-13, verbose = FALSE)
}
\arguments{
\item{pres.pr}{pseudo-reduced pressure}

\item{temp.pr}{pseudo-reduced temperature}

\item{tolerance}{controls the iteration accuracy}

\item{verbose}{print internal}
}
\description{
Hall-Yarborough correlation
}
