% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_tables.R
\name{HY.genDataset7p4t}
\alias{HY.genDataset7p4t}
\title{Generate a dataset of z values read from Standing-Kats chart}
\usage{
HY.genDataset7p4t(to_disk = FALSE)
}
\description{
Generate a dataset of z values read from Standing-Kats chart
}
