% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isValid_correlation}
\alias{isValid_correlation}
\title{Check if supplied correlation (three letter) is valid}
\usage{
isValid_correlation(correlation)
}
\arguments{
\item{correlation}{a z-factor correlation}
}
\description{
Check if supplied correlation (three letter) is valid
}
