% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{multiplotStandingKatz}
\alias{multiplotStandingKatz}
\title{Plot multiple Tpr curves in one figure}
\usage{
multiplotStandingKatz(tpr, pprRange = "lp", ...)
}
\arguments{
\item{tpr}{Pseudo-reduced temperature curve in SK chart}

\item{pprRange}{Takes one of two values: "lp": low pressure, or "hp" for}

\item{...}{additional parameters}
}
\description{
Plot multiple Tpr curves in one figure
}
