% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Papay.R
\name{z.Papay}
\alias{z.Papay}
\title{Papay correlation}
\usage{
z.Papay(pres.pr, temp.pr, tolerance = 1e-13, verbose = FALSE)
}
\arguments{
\item{pres.pr}{pseudo-reduced pressure}

\item{temp.pr}{pseudo-reduced temperature}

\item{tolerance}{rounding tolerance to avoid rounding readings that are in
the middle of the grid. "tolerance" adds flexibility in deciding point closeness.}

\item{verbose}{print internal}
}
\description{
Calculate the Z factor with the Papay's correlation
}
\examples{
# Example 1
## one single z calculation
z.Papay(pres.pr = 1.5, temp.pr = 2.0)
# Example 2
## calculate z for multiple values of Tpr and Ppr
ppr <- c(0.5, 1.5, 2.5, 3.5, 4.5, 5.5, 6.5)
tpr <- c(1.3, 1.5, 1.7, 2)
z.Papay(pres.pr = ppr, temp.pr = tpr)
}
