% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANN.R
\docType{data}
\name{Ppr_min}
\alias{Ppr_min}
\title{# Correlation Kamyab et al. Created using Artificial Neural Networks (ANN)}
\format{An object of class \code{numeric} of length 1.}
\usage{
Ppr_min
}
\description{
# Correlation Kamyab et al. Created using Artificial Neural Networks (ANN)
}
\keyword{datasets}
