% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shell.R
\name{z.Shell}
\alias{z.Shell}
\title{Shell correlation from Kumar thesis (2005)}
\usage{
z.Shell(pres.pr, temp.pr, tolerance = 1e-13, verbose = FALSE)
}
\arguments{
\item{pres.pr}{pseudo-reduced pressure}

\item{temp.pr}{pseudo-reduced temperature}

\item{tolerance}{controls the iteration accuracy}

\item{verbose}{print internal}
}
\description{
Shell correlation from Kumar thesis (2005)
}
