% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sk_short}
\alias{sk_short}
\title{Hall-Yarborough tidy dataset}
\format{An object of class \code{matrix} with 4 rows and 7 columns.}
\usage{
sk_short
}
\description{
Hall-Yarborough tidy dataset
}
\keyword{datasets}
