% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{stats_of_z.stats}
\alias{stats_of_z.stats}
\title{General statistics of z.stats tables}
\usage{
stats_of_z.stats()
}
\description{
This function summarizes the tables generated by z.stats using
custom provided functions. We get a shorter table with statistics of statistics.
}
