% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tpr_13}
\alias{tpr_13}
\title{Example 35 from McCain chapter 03}
\format{An object of class \code{data.frame} with 36 rows and 5 columns.}
\usage{
tpr_13
}
\description{
Example 35 from McCain chapter 03
}
\keyword{datasets}
