% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tpr_15}
\alias{tpr_15}
\title{Data read from Standing-Katz chart at Tpr = 1.5}
\format{An object of class \code{data.frame} with 37 rows and 5 columns.}
\usage{
tpr_15
}
\description{
Data read from Standing-Katz chart at Tpr = 1.5
}
\keyword{datasets}
