% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tpr_17}
\alias{tpr_17}
\title{Data read from Standing-Katz chart at Tpr = 1.7}
\format{An object of class \code{data.frame} with 40 rows and 5 columns.}
\usage{
tpr_17
}
\description{
Data read from Standing-Katz chart at Tpr = 1.7
}
\keyword{datasets}
