% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tpr_20}
\alias{tpr_20}
\title{Data read from Standing-Katz chart at Tpr = 2.0}
\format{An object of class \code{data.frame} with 37 rows and 5 columns.}
\usage{
tpr_20
}
\description{
Data read from Standing-Katz chart at Tpr = 2.0
}
\keyword{datasets}
