% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tpr_240}
\alias{tpr_240}
\title{Data read from Standing-Katz chart at Tpr = 2.4}
\format{An object of class \code{data.frame} with 33 rows and 5 columns.}
\usage{
tpr_240
}
\description{
Data read from Standing-Katz chart at Tpr = 2.4
}
\keyword{datasets}
