% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{z.stats_quantile}
\alias{z.stats_quantile}
\title{Quantiles for z.stats}
\usage{
z.stats_quantile(stat = "MAPE")
}
\arguments{
\item{stat}{Any of the statistical variables in z.stats:
RMSE, MPRE, MAPE, MSE, RSS, MAE}
}
\description{
Calculate the quantiles for any of the statistical variables
that originates from calling z.stats
}
