% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{z_HY}
\alias{z_HY}
\title{Hall-Yarborough tidy dataset}
\format{An object of class \code{data.frame} with 28 rows and 5 columns.}
\usage{
z_HY
}
\description{
28 observations of 5 variables
}
\keyword{datasets}
