% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{z_hy_deriv}
\alias{z_hy_deriv}
\title{Hall-Yarborough tidy dataset}
\format{An object of class \code{matrix} with 4 rows and 7 columns.}
\usage{
z_hy_deriv
}
\description{
Hall-Yarborough tidy dataset
}
\keyword{datasets}
