% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zFactor-package.r
\docType{package}
\name{zFactor-package}
\alias{zFactor-package}
\title{Calculation of compressibility factors using varius correlations}
\description{
Calculation of compressibility factors using varius correlations
}
\details{
zFactor.
}
\author{
Alfonso R. Reyes  \email{alfonso.reyes@oilgainsanalytics.com}
}
