% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Beggs-Brill.R
\name{z.BeggsBrill}
\alias{z.BeggsBrill}
\title{Beggs and Brill correlation}
\usage{
z.BeggsBrill(pres.pr, temp.pr, tolerance = 1e-13, verbose = FALSE)
}
\arguments{
\item{pres.pr}{pseudo-reduced pressure}

\item{temp.pr}{pseudo-reduced temperature}

\item{tolerance}{rounding tolerance to avoid rounding readings that are in
the middle of the grid. "tolerance" adds flexibility in deciding point closeness.}

\item{verbose}{print internal}
}
\description{
Calculate the Z factor with the Brill-Beggs correlation
}
