% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convertStringToVector}
\alias{convertStringToVector}
\title{split a long string to create a vector for testing}
\usage{
convertStringToVector(str)
}
\arguments{
\item{str}{a contnuous long string to split as a vector}
}
\description{
split a long string to create a vector for testing
}
