% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createTidyFromMatrix}
\alias{createTidyFromMatrix}
\title{Create a tidy table from Ppr and Tpr vectors}
\usage{
createTidyFromMatrix(ppr_vector, tpr_vector, correlation = "HY")
}
\arguments{
\item{ppr_vector}{a pseudo-reduced pressure vector}

\item{tpr_vector}{a pseudo-reduced temperature vector}

\item{correlation}{a z-factor correlation}
}
\description{
Create a tidy table from Ppr and Tpr vectors
}
