% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dranchuk-AbouKassem.R
\name{z.DranchukAbuKassem}
\alias{z.DranchukAbuKassem}
\title{Dranchuk- correlation}
\usage{
z.DranchukAbuKassem(pres.pr, temp.pr, tolerance = 1e-13, verbose = FALSE)
}
\arguments{
\item{pres.pr}{pseudo-reduced pressure}

\item{temp.pr}{pseudo-reduced temperature}

\item{tolerance}{controls the iteration accuracy}

\item{verbose}{print internal calclulations}
}
\description{
Dranchuk- correlation
}
