% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dranchuk-Purvis-Robinson.R
\name{z.DranchukPurvisRobinson}
\alias{z.DranchukPurvisRobinson}
\title{Dranchuk-Purvis-Robinson correlation}
\usage{
z.DranchukPurvisRobinson(pres.pr, temp.pr, tolerance = 1e-13,
  verbose = FALSE)
}
\arguments{
\item{pres.pr}{pseudo-reduced pressure}

\item{temp.pr}{pseudo-reduced temperature}

\item{tolerance}{controls the iteration accuracy}

\item{verbose}{print internal}
}
\description{
Dranchuk-Purvis-Robinson correlation
}
