% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_z_correlations}
\alias{get_z_correlations}
\title{Get correlation information}
\usage{
get_z_correlations(how = "short")
}
\arguments{
\item{how}{short: abbreviations; long: description; function: the name of the
correlation function}
}
\description{
Get correlation information
}
