% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Standing-Katz.R
\name{getStandingKatzData}
\alias{getStandingKatzData}
\title{Read file with readings from Standing-Katz chart. Get only the data}
\usage{
getStandingKatzData(tpr = 1.3, pprRange = "lp")
}
\arguments{
\item{tpr}{Pseudo-reduced temperature curve in SK chart}

\item{pprRange}{Takes one of two values: "lp": low pressure, or "hp" for
high pressure}
}
\description{
Read a .txt file that was created from readings of the Standing-Katz chart
and retrieve the points
}
