% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Standing-Katz.R
\name{listStandingKatzCurves}
\alias{listStandingKatzCurves}
\title{List all Standing-Katz curve files available at Low and High pressures}
\usage{
listStandingKatzCurves(pprRange = "lp")
}
\arguments{
\item{pprRange}{Takes one of three values: "lp": low pressure, or "hp" for
high pressure, or 'all' for all the curve text files. The text files reside
under extdata. High pressure is considered above a Ppr > 8.}
}
\description{
List all Standing-Katz curve files available at Low and High pressures
}
\examples{
listStandingKatzCurves(pprRange = 'all')  # list all curves
listStandingKatzCurves(pprRange = 'lp')   # list all the Tpr for low-pressure
#' listStandingKatzCurves(pprRange = 'hp')   # list all the Tpr for high-pressure
}
