% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_tables.R
\name{SK.genDataset7p4t}
\alias{SK.genDataset7p4t}
\title{Generate a dataset of z values read from Standing-Kats chart}
\usage{
SK.genDataset7p4t(to_disk = FALSE)
}
\arguments{
\item{to_disk}{logical indicator to save Rda file to disk. Default FALSE}
}
\description{
Generate a dataset of z values read from Standing-Kats chart
}
