% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{stats_of_z.stats}
\alias{stats_of_z.stats}
\title{General statistics of z.stats tables}
\usage{
stats_of_z.stats(stat = "MAPE")
}
\arguments{
\item{stat}{Any of the statistical variables in z.stats:
RMSE, MPRE, MAPE, MSE, RSS, MAE}
}
\description{
This function summarizes the tables generated by z.stats using
custom provided functions. We get a shorter table with statistics of statistics.
}
\examples{
# Get a statistical summary of the Mean Absolute Percentage Error (MAPE)
stats_of_z.stats()
}
