% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{z_sk_chart}
\alias{z_sk_chart}
\title{Hal;l-Yarborough tidy dataset}
\format{An object of class \code{matrix} with 4 rows and 7 columns.}
\usage{
z_sk_chart
}
\description{
Hal;l-Yarborough tidy dataset
}
\keyword{datasets}
